9/27/2018 Polling Intervals
Correct polling of each address in list for a specified amount of time at a specified interval length between run through: right now when you sleep it stops for m +2 or 3 seconds in between poll sections, the poll sections happen with roughly 1 or 2 seconds in between each address. --> want to make it so these numbers are something you can give the program before it starts running.